runtime: introduce an atfork hook in the runtime
This commit introduce an atfork hook in the runtime, as a follow up to the issue encountered in MPR #455 and MPR #471.

To handle all the specific fork scenarios allowed in Multicore, a hook is required in order to let specific "alterations"
to the runtime do their job: by default, on a single domain program, only the domain lock needs to be reset.
When systhreads are in use, more cleaning up duty is required.

As a result, the usage of pthreads_atfork is also removed, opting to rely in both case on manually calling the
runtime hook when required.