testsuite/lib-random/parallel: spawn many domains
If the number of domains is smaller than the number of cores, and each
domain spawned runs a short computation, I couldn't easily observe
non-deterministic scheduling effects: the domain computations would
happen exactly in the order they are spawned, without actual
parallelism. (In particular, a *broken* implementation of DLS
inheritance would still pass the test.)

This commit changes the test to spawn an extremely-high number of
domains, to ensure that many of them block each other, and
non-deterministic scheduling effects happen.