Reimplement `caml_alloc_small` like in OCaml 4.x
The following idiom is here to stay (it's documented in the OCaml/C API, even though it's in the "low-level interface" section):

    v = caml_alloc_small(sz, tag);
    Field(v, 0) = ... ;
    Field(v, 1) = ... ;

At some point in the development of Multicore OCaml, `caml_alloc_small`
was made to just call into `caml_alloc`.

If `sz <= Max_young_wosize`, as it should, everything works fine; it's just a bit less efficient than in OCaml 4.x, owing to the pre-initialization of the fields of the new value.

However, if the programmer made a mistake and `sz` is bigger than `Max_young_wosize`, the OCaml 4.x implementation in debug mode catches the problem with an assertion, while the `caml_alloc` implementation says nothing and returns a block allocated in the major heap.  Then, the direct assignments `Field(v, i) = ...` mess up the GC invariants, but this will not be caught easily.

Hence: the OCaml 4.x implementation of `caml_alloc_small` is preferable, and this PR brings it back.