fix pthread name setting on FreeBSD/OpenBSD/NetBSD
There is an unfortunate portability mess around setting pthread
names, so we now use the various permutations on FreeBSD
and OpenBSD (which require another header include as well) and
NetBSD (which is similar to Linux, but with a different error
code return).

The code previously used GNU_SOURCE to detect Linux, which isnt
much use as GNU_SOURCE is unconditionally defined at the top
of unix.c, so we just assume that the 'default' platform is
Linux.  pthread_setname_np is only defined in glibc 2.12 or
higher, but looking for that specifically makes it complex to
also detect musl, so it's simplest to leave it as the default
case.